Clean up setViewControllers:
No longer checking to see if viewControllers is different from existing. Also no longer assigning the delegate automatically if the new detail VC is an instance of BDBDetailViewController due to the fact that it could be nested inside a navigation controller. The viewDidLoad method of BDBDetailViewController will now take care of automatically assigning the delegate.